Agregue el primer form request

+++ app/Http/Controllers/ServiceController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Resources\ServiceResource;
 use App\Models\Service;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Validator;
 use Symfony\Component\HttpFoundation\Response;
 
 class ServiceController extends Controller
@@ -14,7 +15,7 @@ class ServiceController extends Controller
      */
     public function index()
     {
-        return response()->json(Service::all(),Response::HTTP_OK);
+        return response()->json(Service::all(), Response::HTTP_OK);
     }
 
     /**
@@ -22,18 +23,26 @@ class ServiceController extends Controller
      */
     public function store(Request $request)
     {
-        $validatedData = $request->validate([
+        $validate = Validator::make($request->all(), [
             'name' => ['required', 'string', 'max:255'],
             'description' => ['sometimes', 'string', 'max:500'],
             'price' => ['required', 'decimal:2'],
         ]);
+
+        if ($validate->fails()) {
+            return response()->json([
+                "error" => $validate->errors()
+            ], Response::HTTP_BAD_REQUEST);
+        }
+
+        $validatedData = $request->all();
 
         $service = Service::create($validatedData);
 
         return response()->json([
             "message" => "Servicio creado con exito",
             "service" => $service
-        ],Response::HTTP_OK);
+        ], Response::HTTP_OK);
     }
 
     /**
@@ -43,7 +52,7 @@ class ServiceController extends Controller
     {
         $service = Service::find($id);
 
-        if($service){
+        if ($service) {
             return response()->json(new ServiceResource($service), Response::HTTP_OK);
         }
 
@@ -56,22 +65,28 @@ class ServiceController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $validatedData = $request->validate([
+        $validate = Validator::make($request->all(), [
             'name' => ['required', 'string', 'max:255'],
             'description' => ['sometimes', 'string', 'max:500'],
             'price' => ['required', 'decimal:2'],
         ]);
 
+        if ($validate->fails()) {
+            return response()->json([
+                "error" => $validate->errors()
+            ], Response::HTTP_BAD_REQUEST);
+        }
+
         $service = Service::find($id);
 
-        if(!$service){
+        if (!$service) {
             return response()->json([
                 "message" => "No existe el servicio con id: {$id}"
             ], Response::HTTP_BAD_REQUEST);
         }
-        
-        $service->update($validatedData);
-        
+
+        $service->update($request->all());
+
         return response()->json([
             "message" => "Servicio actualizado con exito",
             "service" => new ServiceResource($service)
@@ -81,15 +96,23 @@ class ServiceController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy($id)
+    public function destroy($id, $force = false)
     {
         $service = Service::find($id);
 
-        if(!$service){
+        if (!$service) {
             return response()->json([
                 "message" => "No existe el servicio con id: {$id}"
             ], Response::HTTP_BAD_REQUEST);
         }
+
+        if ((sizeof($service->clients) > 0) && !$force) {
+            return response()->json([
+                "message" => "Algunos clientes tienen el servicio registrado"
+            ], Response::HTTP_BAD_REQUEST);
+        }
+
+        $service->clients()->detach();
 
         $service->delete();
